Move phantom_shim into test folder
